refactor: MovieListSection 컴포넌트 파일을 TS로 마이그레이션
